refactor: OrderTableValidator 생성하여 비즈니스 검증 로직 위임
