+++ src/test/java/kitchenpos/ordertable/application/OrderTableServiceTest.java
@@ -4,15 +4,12 @@ import static kitchenpos.common.fixtures.OrderTableFixtures.ORDER_TABLE1_CHANGE_
 import static kitchenpos.common.fixtures.OrderTableFixtures.ORDER_TABLE1_CHANGE_GUEST_REQUEST;
 import static kitchenpos.common.fixtures.OrderTableFixtures.ORDER_TABLE1_CREATE_REQUEST;
 import static kitchenpos.common.fixtures.OrderTableFixtures.ORDER_TABLE1_NUMBER_OF_GUESTS;
-import static kitchenpos.common.fixtures.TableGroupFixtures.TABLE_GROUP1;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.assertj.core.api.SoftAssertions.assertSoftly;
 
 import kitchenpos.common.ServiceTest;
 import kitchenpos.common.fixtures.OrderTableFixtures;
-import kitchenpos.order.OrderStatus;
-import kitchenpos.order.domain.Order;
 import kitchenpos.order.domain.OrderRepository;
 import kitchenpos.ordertable.application.dto.OrderTableChangeEmptyRequest;
 import kitchenpos.ordertable.application.dto.OrderTableChangeGuestRequest;
@@ -21,8 +18,6 @@ import kitchenpos.ordertable.application.dto.OrderTableResponse;
 import kitchenpos.ordertable.domain.OrderTable;
 import kitchenpos.ordertable.domain.OrderTableRepository;
 import kitchenpos.ordertable.exception.OrderTableException;
-import kitchenpos.ordertable.exception.OrderTableException.CannotChangeEmptyStateByOrderStatusException;
-import kitchenpos.tablegroup.domain.TableGroup;
 import kitchenpos.tablegroup.domain.TableGroupRepository;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
@@ -96,43 +91,6 @@ class OrderTableServiceTest extends ServiceTest {
             assertThatThrownBy(() -> orderTableService.changeEmpty(notExistOrderTableId, request))
                     .isInstanceOf(OrderTableException.NotFoundOrderTableException.class)
                     .hasMessage("[ERROR] 해당하는 OrderTable이 존재하지 않습니다.");
-        }
-
-        @Test
-        @DisplayName("주문 테이블의 테이블 그룹이 존재하면 예외가 발생한다.")
-        void throws_ExistTableGroup() {
-            // given
-            final OrderTableChangeEmptyRequest request = ORDER_TABLE1_CHANGE_EMPTY_REQUEST();
-            final OrderTable orderTable = OrderTableFixtures.ORDER_TABLE1();
-            final TableGroup tableGroup = TABLE_GROUP1();
-            final TableGroup savedTableGroup = tableGroupRepository.save(tableGroup);
-            orderTable.updateTableGroupId(savedTableGroup.getId());
-            final OrderTable savedOrderTable = orderTableRepository.save(orderTable);
-
-            // when & then
-            assertThatThrownBy(() -> orderTableService.changeEmpty(savedOrderTable.getId(), request))
-                    .isInstanceOf(OrderTableException.AlreadyExistTableGroupException.class)
-                    .hasMessage("[ERROR] 이미 Table Group이 존재합니다.");
-        }
-
-        @Test
-        @DisplayName("주문 테이블 ID에 해당하는 주문이 존재하고 주문 상태가 조리 or 식사면 예외가 발생한다.")
-        void throws_existsByOrderTableIdAndOrderStatusIn() {
-            // given
-            final OrderTableChangeEmptyRequest request = ORDER_TABLE1_CHANGE_EMPTY_REQUEST();
-            final OrderTable orderTable = new OrderTable(ORDER_TABLE1_NUMBER_OF_GUESTS, false);
-
-            final OrderTable savedOrderTable1 = orderTableRepository.save(orderTable);
-            final int orderLineItemSize = 1;
-
-            final Order order = Order.from(savedOrderTable1.getId(), orderLineItemSize, orderLineItemSize);
-            order.changeStatus(OrderStatus.MEAL);
-            orderRepository.save(order);
-
-            // when & then
-            assertThatThrownBy(() -> orderTableService.changeEmpty(orderTable.getId(), request))
-                    .isInstanceOf(CannotChangeEmptyStateByOrderStatusException.class)
-                    .hasMessage("[ERROR] 주문 테이블의 주문 상태가 조리중이거나 식사중일 때 주문 테이블의 상태를 변경할 수 없습니다.");
         }
     }
